[feat] Task1, DocumentEditor for tool outputs
