+++ tools/components/DocumentEditor/DocumentEditor.jsx
@@ -1,10 +1,10 @@
-'use client';
-
 import React from 'react';
 
-import { PlateEditor } from './editor/plate-editor';
+import { PlateEditor } from './editor/PlateEditor';
 
-const DocumentEditor = ({ markdownContent }) => {
+const DocumentEditor = (props) => {
+  const { markdownContent } = props;
+  
   return (
     <div className="document-editor h-full w-full bg-gray-900 text-white p-6 rounded-lg shadow-lg">
       <div className="bg-gray-800 p-4 rounded-md shadow-md min-h-[400px] border border-gray-700">
