+++ tools/components/DocumentEditor/plate-ui/blockquote-element-static.jsx
@@ -3,17 +3,14 @@ import React from 'react';
 import { cn } from '@udecode/cn';
 import { SlateElement } from '@udecode/plate';
 
-export const BlockquoteElementStatic = ({
-  children,
-  className,
-  ...props
-}) => {
+export const BlockquoteElementStatic = ({ children, className, ...props }) => {
   return (
-    (<SlateElement
+    <SlateElement
       as="blockquote"
       className={cn(className, 'my-1 border-l-2 pl-6 italic')}
-      {...props}>
+      {...props}
+    >
       {children}
-    </SlateElement>)
+    </SlateElement>
   );
 };
