+++ tools/components/DocumentEditor/plate-ui/editor.jsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 
 import { cn } from '@udecode/cn';
@@ -7,8 +9,6 @@ import {
   useEditorRef,
 } from '@udecode/plate/react';
 import { cva } from 'class-variance-authority';
-
-('use client');
 
 const editorContainerVariants = cva(
   'relative w-full cursor-text select-text overflow-y-auto caret-primary selection:bg-brand/25 focus-visible:outline-none [&_.slate-selection-area]:z-50 [&_.slate-selection-area]:border [&_.slate-selection-area]:border-brand/25 [&_.slate-selection-area]:bg-brand/15',
